requirements.txt
streamlit
pydantic
pytesseract
pdf2image
Pillow
plotly
pandas
numpy
